Refactoring code following the discovery of an unused dummy circulation API.
